wip Squashed some bugs.
Now the pin edge ISR is not used. Just poll. There is no setting for
rising/falling interrupts combined, which is what I need.
 
Need to debug connection on LH_DOOR signal, it's not making it to the
hex buffer. Seems like it gets to D5 cathode, but not to the buffer.
Also, I can't find L6.